# packages/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/lines": {
            "get": {
                "description": "Retrieves information about all available transit lines.\nThis endpoint returns a list of all transit lines and their associated stations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lines"
                ],
                "summary": "Get all lines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all lines data.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Not Modified: The ETag matches the If-None-Match header."
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Error retrieving lines data."
                    }
                }
            }
        },
        "/lines/metadata": {
            "get": {
                "description": "Retrieves the colour and mode of every line in the network.\nClients use this instead of deriving either from the line name, which is a per city convention rather than a fact.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lines"
                ],
                "summary": "Get the metadata of all lines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the line metadata.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/utils.LineMetadataEntry"
                            }
                        }
                    },
                    "304": {
                        "description": "Not Modified: The ETag matches the If-None-Match header."
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lines/segments": {
            "get": {
                "description": "Retrieves information about all available transit segments.\nA segment is the geojson lineString part of a line between two stations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lines"
                ],
                "summary": "Get all segments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GeoJSON segments data",
                        "schema": {
                            "$ref": "#/definitions/utils.SegmentsCollection"
                        }
                    },
                    "304": {
                        "description": "Not Modified"
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Error retrieving segments data.",
                        "schema": {}
                    }
                }
            }
        },
        "/lines/{lineId}/{stationId}/statistics": {
            "get": {
                "description": "Retrieves statistics for a specific line at a specific station.\nThis endpoint returns the number of reports for the specified line at the given station within the given time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lines"
                ],
                "summary": "Get line statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the line",
                        "name": "lineId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the station",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time for the statistics (format: RFC3339)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time for the statistics (format: RFC3339)",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved line statistics.",
                        "schema": {
                            "$ref": "#/definitions/statistics.Statistics"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid time format.",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity: The station or line does not belong to the network.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Failed to get number of reports.",
                        "schema": {}
                    }
                }
            }
        },
        "/lines/{lineName}": {
            "get": {
                "description": "Retrieves information about a specific transit line.\nThis endpoint returns the details of a single line, including all its stations, based on the provided line name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lines"
                ],
                "summary": "Get a single line",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the line (e.g., S1, U2)",
                        "name": "lineName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the specified line data.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Line not found: The specified line does not exist.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Error retrieving line data."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "statistics.Statistics": {
            "type": "object",
            "properties": {
                "numberOfReports": {
                    "type": "integer"
                }
            }
        },
...
        "utils.GeoJSONCRS": {
            "type": "object",
            "properties": {
                "properties": {
                    "type": "object",
                    "additionalProperties": true
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "utils.LineMetadataEntry": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "isCircular": {
                    "type": "boolean"
                },
                "mode": {
                    "type": "string"
                }
            }
        },
...
        "utils.SegmentFeature": {
            "type": "object",
            "properties": {
                "geometry": {
                    "$ref": "#/definitions/utils.SegmentGeometry"
                },
                "properties": {
                    "$ref": "#/definitions/utils.SegmentProperties"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "utils.SegmentGeometry": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "utils.SegmentProperties": {
            "type": "object",
            "properties": {
                "line_color": {
                    "type": "string"
                },
                "sid": {
                    "type": "string"
                }
            }
        },
        "utils.SegmentsCollection": {
            "type": "object",
            "properties": {
                "crs": {
                    "$ref": "#/definitions/utils.GeoJSONCRS"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.SegmentFeature"
                    }
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}